DataFrame: rename files to match primary export
