+++ packages/grafana-data/src/utils/index.ts
@@ -12,8 +12,9 @@ export * from './object';
 export * from './moment_wrapper';
 export * from './thresholds';
 export * from './text';
-export * from './dataFrameHelper';
-export * from './dataFrameView';
+export * from './MutableDataFrame';
+export * from './FieldCache';
+export * from './DataFrameView';
 export * from './vector';
 
 export { getMappedValue } from './valueMappings';
